+++ cli/image_describe.go
@@ -48,5 +48,14 @@ func runImageDescribe(cli *CLI, cmd *cobra.Command, args []string) error {
 	fmt.Printf("Protection:\n")
 	fmt.Printf("  Delete:\t%s\n", yesno(image.Protection.Delete))
 
+	fmt.Print("Labels:\n")
+	if len(image.Labels) == 0 {
+		fmt.Print("  No labels\n")
+	} else {
+		for key, value := range image.Labels {
+			fmt.Printf("  %s: %s\n", key, value)
+		}
+	}
+
 	return nil
 }